Extract FileCompiler to its file
